Move transactions UI components to transactions/ folder
