+++ packages/client/src/transactions/CosmosTransactionDetail.tsx
